Fix of large class code smell on the class SessionServiceController.java who belong to the packege Web
